Players track their position.
This is redundant - player position is already captured in the
model elsewhere, but I think it'll keep the update code cleaner
since we won't have to loop through the whole path to find a
player when we want to move them.  We'll have to make sure to
keep the two numbers in sync, but that should be doable.